+++ plugins/login/classes/RememberMe/RememberMe.php
@@ -4,6 +4,13 @@ namespace Grav\Plugin\Login\RememberMe;
 
 use Birke\Rememberme\Authenticator;
 
+/**
+ * RememberMe
+ *
+ * Handles persistent cookie-storage (Remember Me)
+ *
+ * @author  Sommerregen <sommerregen@benjamin-regler.de>
+ */
 class RememberMe extends Authenticator
 {
     /**