+++ plugins/login/CHANGELOG.md
@@ -1,4 +1,19 @@
-# v1.01
+# v1.1.0
+## 12/01/2015
+
+1. [](#new)
+    * Support new **User Registration**
+1. [](#improved)
+    * Use new security salt for newer and fallback otherwise
+    * Composer update of libraries
+    * Check for session existence else throw a runtime error
+1. [](#bugfix)
+    * Fix remember-me functionality
+    * Check page exists so as not to fail hard
+    * Fix for static Inflector references #17
+
+    
+# v1.0.1
 ## 11/23/2015
 
 1. [](#improved)